# swagger.json
{
  "swagger": "2.0",
  "info": {
    "description": "",   
    "version": "1.0.2",
    "title": "testdb",
    "termsOfService": "",
    "contact": {
      "email": "shubhamjulaha.rexweb@gmail.com"
    },
    "license": {
      "name": "",
      "url": ""
    }
  },
  "host": "192.168.43.50:3000",
  "basePath": "",
...
  "schemes": [
    "http"
  ],
  "paths": {
...
    "/api/proposalsubmission": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Proposal Submission Module"
        ],
        "summary": "This API will provide list of all Proposal Submission data along with client information.",
        "description": "All proposal submission fields",
        "operationId": "getproposalsubmission",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      },
      "post": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Proposal Submission Module"
        ],
        "summary": "This API is for Adding a new Proposal Submission entry and client record into the database ",
        "description": "Add a new field into db",
        "operationId": "addproposalsubmission",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/ProposalSubmission"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
    "/api/proposalsubmission/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Proposal Submission Module"
        ],
        "summary": "This API will display proposal submission table entry data by Id",
        "description": "List of Proposal Submission Module on the basis of ID",
        "operationId": "getProposalSubmissionById Matrix",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of Proposal Submission field  for getting department nnfo by ID .",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      },
      "patch": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Proposal Submission Module"
        ],
        "summary": "This API will update Proposal Submission Module entries on the basis of ID",
        "description": "Update  fields information into db",
        "operationId": "patchProposalSubmissionModule",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID for updating Proposal submission table fields  ",
            "required": true,
            "type": "integer"
          },
          {
            "in": "body",
            "name": "body",
            "description": "",
            "required": true,
            "schema": {
              "$ref": "#/definitions/EditProposalSubmission"
            }
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      },
      "delete": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Proposal Submission Module"
        ],
        "summary": "This API will Set (isActive=0) in Proposal Submission Module on the baisi of ID",
        "description": "",
        "operationId": "deleteProposalSubmission",
        "produces": [
          "application/xml",
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of Proposal Submission Module that needs to be deleted",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Please fill all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
    "/api/proposalsubmission/?userId=13&portal=upwork": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Proposal Submission Module"
        ],
        "summary": "This API will display list of today proposals having portal upwork ",
        "description": "list of today proposals having portal upwork",
        "operationId": "getProposallistByuserId Matrix",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
    "/api/discussion/{id}": {
      "get": {
        "security": [
          {
            "Bearer": []
          }
        ],
        "tags": [
          "Discussion Module"
        ],
        "summary": "This API will get leads record on the basis of Id",
        "description": "Return client information fields on the basis of id",
        "operationId": "getdiscussionfields",
        "consumes": [
          "application/json"
        ],
        "produces": [
          "application/json"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "description": "ID of the field for getting information.",
            "required": true,
            "type": "integer"
          }
        ],
        "responses": {
          "400": {
            "description": "Please FIll all fields correctly"
          },
          "200": {
            "description": "Success with data"
          }
        }
      }
    },
...
  },
  "securityDefinitions": {
    "Bearer": {
      "type": "apiKey",
      "name": "Authorization",
      "in": "header"
    }
  },
  "definitions": {
...
    "ProposalSubmission": {
      "type": "object",
      "properties": {
        "profile": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "portal": {
          "type": "string"
        },
        "clientName": {
          "type": "string"
        },
        "countryCityState": {
          "type": "string"
        },
        "company": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "integer"
        },
        "domain": {
          "type": "string"
        },
        "submissionTitle": {
          "type": "string"
        },
        "pitchContent": {
          "type": "string"
        },
        "jobReqLink": {
          "type": "string"
        },
        "proposalLink": {
          "type": "string"
        },
        "clientPostedDateTime": {
          "type": "string"
        },
        "projectType": {
          "type": "string"
        },
        "TargetAchieved": {
          "type": "integer"
        },
        "proposalSubmissionDateTime": {
          "type": "string"
        },
        "isActive": {
          "type": "integer"
        }
      }
    },
    "EditProposalSubmission": {
      "type": "object",
      "properties": {
        "profile": {
          "type": "string"
        },
        "date": {
          "type": "string"
        },
        "portal": {
          "type": "string"
        },
        "clientName": {
          "type": "string"
        },
        "countryCityState": {
          "type": "string"
        },
        "company": {
          "type": "string"
        },
        "email": {
          "type": "string"
        },
        "phone": {
          "type": "integer"
        },
        "domain": {
          "type": "string"
        },
        "submissionTitle": {
          "type": "string"
        },
        "pitchContent": {
          "type": "string"
        },
        "jobReqLink": {
          "type": "string"
        },
        "proposalLink": {
          "type": "string"
        },
        "clientPostedDateTime": {
          "type": "string"
        },
        "projectType": {
          "type": "string"
        },
        "TargetAchieved": {
          "type": "integer"
        },
        "proposalSubmissionDateTime": {
          "type": "string"
        },
        "status": {
          "type": "string"
        }
      }
    },
...
  },
...
}